Cleanup diagnostics
